Started refactoring regexp impl to use gnu-regexp
